requirements.txt
alembic==0.7.0
cssmin==0.2.0
DateTime==4.0.1
fake-factory==0.5.3
fixtures==3.0.0
Flask==0.10.1
Flask-Assets==0.10
Flask-Bootstrap==3.3.0.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Migrate==1.3.0
Flask-Mobility==0.1.1
Flask-Redis==0.1.0
Flask-RESTful==0.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-Views==0.2.1
Flask-WTF==0.10.3
eventlet==0.18.4
freezegun==0.3.6
funcsigs==0.4
futures==3.0.3
Jinja2==2.8
lea==2.2.0
mysql-connector-python-rf==2.0.4
mock==1.0.1
pep8==1.6.2
pytest==2.9.1
pytest-assume==1.2
pytest-cache==1.0
pytest-cov==1.8.1
pytest-mock==0.4.0
python-dateutil==2.5.2
python-editor==0.5
python-intercom==2.1.1
python-Levenshtein==0.12.0
python-magic==0.4.6
python-mimeparse==1.6.0
python-subunit==1.2.0
python-twitter==2.2
pytz==2015.6
PyYAML==3.11
redis==2.10.5
requests==2.7.0
requests-oauthlib==0.5.0
SQLAlchemy==1.2.19
tzlocal==1.1.2
ua-parser==0.3.6
unittest2==1.1.0
uritemplate==0.6
urllib3==1.12
vcrpy==1.6.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
webassets==0.10.1
Werkzeug==0.10.4
WTForms==2.0.1
